refactor: tighten sqlite backend boundaries
