+++ src/infra/adapters/sqlite/sqlite3/parser/mod.rs
@@ -2,15 +2,15 @@ mod command_tag;
 mod lexer;
 mod output;
 
-pub(in crate::adapters::sqlite) use command_tag::{
+pub(super) use command_tag::{
     aggregate_sqlite_command_tag, command_tag_result, sqlite_statement_tags,
     statement_counts_as_select_tag,
 };
-pub(in crate::adapters::sqlite) use lexer::{
+pub(super) use lexer::{
     append_changes_query, is_sqlite_rerunnable_export_query, sqlite_adhoc_execution_query,
     sqlite_export_not_rerunnable_error, sqlite_probe_marker, try_split_sqlite_statements,
 };
-pub(in crate::adapters::sqlite) use output::{
+pub(super) use output::{
     last_sqlite_result_set, parse_affected_rows, parse_count_result, quoted_to_query_result,
     strip_sqlite_probes,
 };
