+++ src/infra/adapters/sqlite/sqlite3/parser/output.rs
@@ -86,7 +86,7 @@ fn decode_hex_text(hex: &str) -> Result<String, DbOperationError> {
     Ok(String::from_utf8_lossy(&bytes).into_owned())
 }
 
-pub(in crate::adapters::sqlite) fn parse_unistr_inner_sql_escapes(
+pub(in crate::adapters::sqlite::sqlite3) fn parse_unistr_inner_sql_escapes(
     value: &str,
 ) -> Result<String, DbOperationError> {
     let inner = value
@@ -148,7 +148,7 @@ fn decode_sqlite_nul_text_transport(text: &str) -> Result<Option<String>, DbOper
     Ok(None)
 }
 
-pub(in crate::adapters::sqlite) fn decode_preview_transport_unistr(
+pub(in crate::adapters::sqlite::sqlite3) fn decode_preview_transport_unistr(
     value: &str,
 ) -> Result<Option<String>, DbOperationError> {
     let inner = parse_unistr_inner_sql_escapes(value)?;
@@ -173,7 +173,7 @@ fn decode_hex_bytes(hex: &str) -> Result<Vec<u8>, DbOperationError> {
     Ok(bytes)
 }
 
-pub(in crate::adapters::sqlite) fn parse_quoted_value(
+pub(in crate::adapters::sqlite::sqlite3) fn parse_quoted_value(
     value: &str,
     source: QuerySource,
     decode_preview_transport: bool,
@@ -239,7 +239,7 @@ fn parse_quoted_records(
         .collect()
 }
 
-pub(in crate::adapters::sqlite) fn quoted_to_query_result(
+pub(in crate::adapters::sqlite::sqlite3) fn quoted_to_query_result(
     query: &str,
     stdout: &str,
     source: QuerySource,
@@ -281,7 +281,7 @@ pub(in crate::adapters::sqlite) fn quoted_to_query_result(
     ))
 }
 
-pub(in crate::adapters::sqlite) fn last_sqlite_result_set(
+pub(in crate::adapters::sqlite::sqlite3) fn last_sqlite_result_set(
     stdout: &str,
     marker: &str,
 ) -> Result<Option<String>, DbOperationError> {
@@ -334,7 +334,7 @@ pub(in crate::adapters::sqlite) fn last_sqlite_result_set(
     Ok(last_result)
 }
 
-pub(in crate::adapters::sqlite) fn strip_sqlite_probes(
+pub(in crate::adapters::sqlite::sqlite3) fn strip_sqlite_probes(
     stdout: &str,
     marker: &str,
 ) -> Result<(String, HashMap<usize, usize>), DbOperationError> {
@@ -432,7 +432,7 @@ fn last_csv_cell(stdout: &str) -> Result<String, DbOperationError> {
         .ok_or_else(|| DbOperationError::EmptyResponse(stdout.to_string()))
 }
 
-pub(in crate::adapters::sqlite) fn parse_affected_rows(
+pub(in crate::adapters::sqlite::sqlite3) fn parse_affected_rows(
     stdout: &str,
 ) -> Result<usize, DbOperationError> {
     last_csv_cell(stdout)
@@ -446,7 +446,7 @@ pub(in crate::adapters::sqlite) fn parse_affected_rows(
         .map_err(|error| DbOperationError::CommandTagParseFailed(error.to_string()))
 }
 
-pub(in crate::adapters::sqlite) fn parse_count_result(
+pub(in crate::adapters::sqlite::sqlite3) fn parse_count_result(
     stdout: &str,
 ) -> Result<usize, DbOperationError> {
     first_csv_cell(stdout)
